# app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BookLib",
    "version": "0.1.0"
  },
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get a list with all registered users",
        "description": "Returns a list with all registered users on the database, or en empty list\nif there are none.",
        "operationId": "get_all_users_get",
        "responses": {
          "200": {
            "description": "List with all registered users on the database",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/UserDBResponse"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "type": "array",
                  "title": "Response Get All Users Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create new user",
        "description": "Inserts a new user, and returns the created database entity, if the values\nfollow the next criteria:\n\n* The *username* and *email* are unique (no other user uses them already)\n* The *email* is in email format\n* The *password* length is less than or equals to 72 Bytes.",
        "operationId": "create_user_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User created and returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDBResponse"
                }
              }
            }
          },
          "422": {
            "description": "Username or email not unique, or email in incorrect format",
            "content": {
              "application/json": {
                "example": {
                  "detail": "< reason for failure >"
                }
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "User login",
        "description": "Login user with username and password.",
        "operationId": "login_users_login_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDBResponse"
                }
              }
            }
          },
          "401": {
            "description": "Wrong credentials",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Wrong Credentials! Username and/or password are incorrect."
                }
              }
            }
          },
          "422": {
            "description": "Invalid format",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid format of username and/or password"
                }
              }
            }
          }
        }
      }
    },
    "/users/{username}": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update user",
        "description": "Updates the user information if the new values follow the next criteria:\n\n* The *username* and *email* are unique (no other user uses them already)\n* The *email* is in email format\n* The *password* length is less than or equals to 72 Bytes.",
        "operationId": "update_user_users__username__put",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User with updated information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDBResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "There is no user with that username."
                }
              }
            }
          },
          "422": {
            "description": "New user or email already registered",
            "content": {
              "application/json": {
                "example": {
                  "detail": "An user with that username or email already exists."
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete user",
        "description": "Delete user with the given username",
        "operationId": "delete_user_users__username__delete",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "There is no user with that username."
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "UserCreate": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "examples": [
              "username"
            ]
          },
          "password": {
            "type": "string",
            "title": "Password",
            "examples": [
              "password_clear"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "user@example.com"
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "examples": [
              true
            ]
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin",
            "examples": [
              false
            ]
          }
        },
        "type": "object",
        "required": [
          "username",
          "password",
          "email",
          "is_active",
          "is_admin"
        ],
        "title": "UserCreate",
        "description": "Pytdantic model for User creation"
      },
      "UserDBResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "examples": [
              "username"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "user@example.com"
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "examples": [
              true
            ]
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin",
            "examples": [
              false
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "examples": [
              "2026-03-24T16:14:21.554Z"
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "examples": [
              "2026-03-24T16:14:21.554Z"
            ]
          },
          "last_login": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login",
            "examples": [
              "2026-03-24T16:14:21.554Z"
            ]
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "is_active",
          "is_admin",
          "created_at",
          "updated_at",
          "last_login"
        ],
        "title": "UserDBResponse",
        "description": "Pytdantic model for User response"
      },
      "UserLogin": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "examples": [
              "username"
            ]
          },
          "password": {
            "type": "string",
            "title": "Password",
            "examples": [
              "password_clear"
            ]
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserLogin",
        "description": "Pydantic model for User login."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}